# api-docs.yaml
openapi: 3.0.0
paths:
...
  /api/v1/favorites:
    get:
      operationId: FavoritesController_getUserFavorites
      summary: Get user favorites
      parameters: []
      responses:
        '200':
          description: Favorites retrieved successfully
      tags: &ref_7
        - favorites
      security: &ref_8
        - bearer: []
    post:
      operationId: FavoritesController_addFavorite
      summary: Add product to favorites
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddFavoriteDto'
      responses:
        '201':
          description: Product added to favorites
        '404':
          description: Product not found
        '409':
          description: Product already in favorites
      tags: *ref_7
      security: *ref_8
  /api/v1/favorites/{productId}:
    delete:
      operationId: FavoritesController_removeFavorite
      summary: Remove product from favorites
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Product removed from favorites
        '404':
          description: Favorite not found
      tags: *ref_7
      security: *ref_8
  /api/v1/favorites/count/{productId}:
    get:
      operationId: FavoritesController_getFavoriteCount
      summary: Get favorite count for a product
      parameters:
        - name: productId
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Favorite count retrieved
      tags: *ref_7
      security: *ref_8
  /api/v1/favorites/test:
    get:
      operationId: FavoritesController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_7
      security: *ref_8
...
  /api/v1/notifications:
    get:
      operationId: NotificationsController_getNotifications
      summary: Get user notifications
      parameters:
        - name: isRead
          required: true
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Notifications retrieved successfully
      tags: &ref_15
        - notifications
      security: &ref_16
        - bearer: []
  /api/v1/notifications/unread-count:
    get:
      operationId: NotificationsController_getUnreadCount
      summary: Get unread notification count
      parameters: []
      responses:
        '200':
          description: Unread count retrieved successfully
      tags: *ref_15
      security: *ref_16
  /api/v1/notifications/{id}/read:
    put:
      operationId: NotificationsController_markAsRead
      summary: Mark notification as read
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Notification marked as read
        '404':
          description: Notification not found
      tags: *ref_15
      security: *ref_16
  /api/v1/notifications/read-all:
    put:
      operationId: NotificationsController_markAllAsRead
      summary: Mark all notifications as read
      parameters: []
      responses:
        '200':
          description: All notifications marked as read
      tags: *ref_15
      security: *ref_16
  /api/v1/notifications/{id}:
    delete:
      operationId: NotificationsController_delete
      summary: Delete a notification
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Notification deleted successfully
        '404':
          description: Notification not found
      tags: *ref_15
      security: *ref_16
  /api/v1/notifications/test:
    get:
      operationId: NotificationsController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_15
      security: *ref_16
  /api/v1/chat:
    post:
      operationId: ChatController_createChat
      summary: Create a new chat
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChatDto'
      responses:
        '201':
          description: Chat created successfully
        '400':
          description: Invalid request
      tags: &ref_17
        - chat
      security: &ref_18
        - bearer: []
    get:
      operationId: ChatController_getUserChats
      summary: Get all user chats
      parameters: []
      responses:
        '200':
          description: Chats retrieved successfully
      tags: *ref_17
      security: *ref_18
  /api/v1/chat/{id}:
    get:
      operationId: ChatController_getChatById
      summary: Get chat by ID with messages
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Chat retrieved successfully
        '403':
          description: Forbidden
        '404':
          description: Chat not found
      tags: *ref_17
      security: *ref_18
  /api/v1/chat/{id}/block:
    put:
      operationId: ChatController_blockChat
      summary: Block a chat
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Chat blocked successfully
        '403':
          description: Forbidden
      tags: *ref_17
      security: *ref_18
  /api/v1/chat/{id}/unblock:
    put:
      operationId: ChatController_unblockChat
      summary: Unblock a chat
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: Chat unblocked successfully
        '403':
          description: Forbidden
      tags: *ref_17
      security: *ref_18
  /api/v1/chat/test:
    get:
      operationId: ChatController_test
      summary: Test endpoint
      parameters: []
      responses:
        '200':
          description: Test successful
      tags: *ref_17
      security: *ref_18
info:
  title: memi.az API
  description: Recommerce Platform Backend API
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
...
    AddFavoriteDto:
      type: object
      properties: {}
...
    CreateChatDto:
      type: object
      properties: {}