# src/modules/chat/dtos/create-chat.dto.ts
import { IsUUID } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class CreateChatDto {
  @ApiProperty({
    example: '123e4567-e89b-12d3-a456-426614174000',
    description: 'Product ID to start a chat about',
  })
  @IsUUID()
  productId: string;
}