# src/modules/favorites/dtos/add-favorite.dto.ts
import { IsUUID } from 'class-validator';
import { ApiProperty } from '@nestjs/swagger';

export class AddFavoriteDto {
  @ApiProperty({
    example: '123e4567-e89b-12d3-a456-426614174000',
    description: 'Product ID to add to favorites',
  })
  @IsUUID()
  productId: string;
}